animator/renderer bug fixes
git-svn-id: http://source.usc.edu/svn/scec_vdo_vtk@296 25aa0b20-aa2b-420e-afc3-456cccd12aca

+++ trunk/src/org/scec/vtk/timeline/gui/TimelinePanel.java
@@ -437,7 +437,7 @@ implements MouseListener, MouseMotionListener, AnimationTimeListener, TimelinePl
 			double time = getTime(e.getX());
 			if (pluginIndex == INDEX_TAG_HEADER) {
 				// clicked header, go to that time
-				System.out.println("Going to time "+time);
+				System.out.println("Going to time "+(float)time);
 				timeline.activateTime(time);
 			} else if (pluginIndex == INDEX_TAG_CAMERA) {
 				// create a camera key
